backend/requirements.txt
# Backend dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# CORS
python-multipart>=0.0.6

# AI/ML Dependencies (Required for Been There)
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.1.0

# Groq API for chat interface
groq>=0.4.0

# Supabase client
supabase>=2.0.0

# Optional: For enhanced classification
# tensorflow>=2.14.0  # For neural network moderator
